Added woxser to local repo
It turns out that woxser needed to be in the local repo, which makes
sense. In fact, I don't know how I did without it before. I found that
adding a flag to the mvn install command makes it update the repo if
there are problems the first time, so I added this command to the
tutorial.