#25 Able to launch VizDoom from RL4J classes
RL4J has a class that treats VizDoom like an MDP, and now my code can
launch the game through that class. However, I have to change some of
the RL4J code. This seems to mainly be because I am using an older
version of VizDoom. Their API maps to the most up-to-date VizDoom
implementation, but getting VizDoom to work in Java was such a huge
hassle for me that I'm going to stick to my old VizDoom version.

This means I should be able to delete the vizdoom.fromRL4J package and
keep my vizdoom.jar