#25 Learn Doom with DQN
Wiped out my code (which was trying to use the Toy network) and replaced
it with code I found here:
https://github.com/rubenfiszel/rl4j-examples/blob/master/src/main/java/org/deeplearning4j/rl4j/Doom.java

When applied to DeadlyCorridor, it is clear that the agent is running
trials (when I set render true), and it seems to be improving, though
this isn't clear. I'm going to run it overnight with render=false and
see if it finishes, and if I can load the result after the fact (and
also plot performance).

Note that videos seem to be saved to the rl4j-data dir, but they are all
of black screen. I'm assuming this is because I am using my old version
of vizdoom instead of the latest version, but am not sure. Maybe worth
looking into more.