#19 Integrated convolutionDeltas option
Turning this on should define the CPPN inputs for creating the
convolutional network in the same manner as Verbancsics. This seems to
work fine for networks using separate outputs for each layer pairing,
but actually does not work well (in Tetris) when using substrate
coordinate inputs ... but this is exactly what Verbancsics did! The only
difference I can think of that might be having an impact is the use of
the substrateMapping. Verbancsics used center scaling in all
experiments, but my Tetris substrateMapping goes from top to bottom.
Worth checking.