#15 Code now allows all weights in receptive field for one feature to be the same
Originally, I was going to completely remove two of the CPPN inputs (for
target neuron coordinate) to make this work, but that ended up causing
problems. This is because, even in a network that is convolutional,
there will still be some fully connected components ... for example,
connecting the final processing layers to the output(s). In order to
define these links, the CPPN needs to have inputs for both source and
target neurons. However, when defining links in the convolutional
portions, the extra inputs are 0.

Also added a batch file to test all of this.