#25 Model loading works for VizDoom
Thanks to the help I got here
(https://github.com/deeplearning4j/rl4j/issues/61#issuecomment-327925186)
I was able to fix the network/model loading. The problem is that the
play method needs to take in a HistoryProcessor along with the MDP in
order to know how to properly decode the game state.

I'm not certain of this, but thinking back to the breakthough paper
about DQN in VizDoom I think that they fed in 4 frames of screen
information at a time, and this might be what the HistoryProcessor is
telling the DQN to do here, since the HistoryProcessor.Configuration has
a History length of 4.