#27 Simple test of the VGG16 ImageNet
This short program can take an image and assign labels to it based on
the pre-trained image net. Need to dissect this a bit more by finding
out how to take an Image from memory and turn it into an INDArray, and
also need to be able to read all label scores.