#27 Tested on images without preprocessing
This test confirms that the code runs just fine if there is no
preprocessing on an image that is already the correct size. Also, these
tests were done on some CPPN images, and although the results were
weird, they were not unreasonable.

Also removed some old scope art files that we don't need.