Introduced EA class
The EA interface is meant to be a general interface for all EAs, but it contains no method signatures. It just binds the GenerationalEA (formerly the only type of EA) and the SteadyStateEA (new for MAP Elites) to a common type. This required many fixes throughout the code, since many assumed they could call the methods of GenerationalEA. This was fixed by casting MMNEAT.ea wherever necessary, but it made me realize that MMNEAT.ea is probably being directly accessed in too many places.