#27 Better image saving
It was a bit absurd that all image saving went through the DrawingPanel class. I've fixed this with a method in GraphicsUtil which is now used in Picbreeder, the InnovationEngine, and elsewhere.

I also added a main method to the Innovation Engine to make it easier to launch, and tweaked the batch file parameters (still need to test out)