#27 Testing different ImageNet models
Strangely, AlexNet with pre-trained ImageNet weights does not seem to be
supported in DL4J yet. However, I tested with VGG19, and it was slightly
better than VGG16, though still disappointing in comparison with the
labels stated in the Understanding Innovation Engines paper.

Clearly, the image net model should be a command line parameter