#27 Tested with GoogLeNet
I feel that results were generally worse with GoogLeNet, but it got some
labels correct (at least, the same as in the Understanding Innvation
Engines paper) that the other methods did not. This makes me wonder if
they should be combined somehow ... clearly this would be
computationally expensive, but perhaps useful.

Maybe alternating which model is used at each step would be a good idea