#27 Changes MAP Elites code a lot
After re-reading the innovation engines paper, I made several changes to make my approach more like thiers. Primarily, a genome can be stored in many bins now, and it does not have to be stored in the bin which corresponds to its best score. Even if a genome's best score corresponds to one particular bin, it may also be better than the current elite in another bin.

This definitely makes sense for Innovation Engines, but I fear it will make MAP Elites different to use in other domains in the future.

Also, the ImageNet being used now is VGG19 instead of VGG16