#27 Save a final archive, and save all elites
Changed the saving behavior.

First, all xml files for elites must be saved so that resume on failure is possible. However, I made it so that all files overwrite, so I will no longer keep extra xml network files.

Also, if the simulation actually ends, it saves an archive at the end of all the final pictures.