#27 Added support for resuming after crash
The necessary information needed for resuming an aborted simulation is now being saved. This includes the scores for all elites that are saved. Also have to update parameters. Additionally, a log with some useful information is now saved, but I should add a plot file that makes it easy to look at its contents.