#27 Add other ways of using all zoo models
In addition to averaging across all models, I also added a way of taking the min across all models.

I also started adding a way that concatenates the scores of all models, but this is not finished yet because I need to figure out the INDArrays. Also, this approach will require a difference binning strategy for MAP Elites.