#34 Still can't recreate original picbreeder images
I aded a method to sort the node genes so all links are feedforward, and
also checked the different output neurons, but the results are still
disappointing. There aren't just wrong, they are incredibly bland, which
is especially strange for such complicated networks, which makes me
think there is some scaling issue with how the final outputs are
interpreted.

Note that I also have to tweak the XML files for any genome I load to
prevent an error when trying to load.