#24 Steps toward a DL4J agent in Tetris
Working on class for experiment RL4JExperiment.

Figured out that if I apply DL4J to HyperNEATTasks, then I can use substrate information to determine the shape of the input/output volumes. I might even be able to use substrate information to define other aspects of the network architecture, but I'm not fully there yet.

There is still work to be done in TensorNetworkFromHyperNEATSpecification.

Also: if this plan works, then I will still only be able to evaluate DL4J networks, but I'm not sure where/when weights need to be updated for learning. Perhaps also inside TensorNetworkFromHyperNEATSpecification?