#24 The TensorNetworkFromHyperNEATSpecification improving
The TensorNetworkFromHyperNEATSpecification class constructs a network, and is sytactically able to be applied to a domain like Tetris. However, I need to customize the initialization steps to take the substrate configuration into account.

Also need a way of launching an RL experiment in general, and to also work on when learning/backprop occurs.