#24 Runs, but crashes
I've set up TensorNetworkFromHyperNEATSpecification and RL4JExperiment
to the point where they load the domain, create a network, and try to
play Tetris. However, I immediately get an exception. The key issue
seems to be that the network is expecting a "batch" of training examples
rather than a single input example, though there may be other issues
beyond this.