#34 Tweak in interpretation of HSB values
Based on email conversation with Joost, I slightly changed my
interpretation of the HSB values when generating images. I don't think
this change was strictly necessary, but I've tested it out in
Picbreeder, and the images still look good, so I'm going to keep this
convention to stay consistent.

However, some additional scaling is needed to be fully compatible with
the original Picbreeder genotypes