#24 Can run CartPole once
On the path to getting Tetris to work, I took a massive step backward to
try to get the RL Glue version of CartPole to work. Specifically, I
wanted to do it within the RL4J framework, using a DQN. This allowed me
to copy a lot of RL4J code, but made me realize that I can't simply send
a DL4J network into my Task instances. Rather, I have to make my task
instances implement RL4J's MDP interface.

I've done this for RLGlueTask, and changed the code in RL4JExperiment to
acommodate this change (it is kind of a mess though). Note that this
approach will not work with Tetris as is, since even though Tetris is an
RLGlueTask, it relies on the TetrisAfterStateAgent.

Also, even though RL4JExperiment now runs the RL Glue CartPole task with
the RL4J framework, it only successds once ... after that, it seems to
not reset properly. Will need to look into it more.