#24 Basic RL Glue domains run fine now, but learning sucks
The bug with only running one episode and then failing to reset was
actually easy to fix, and had to do with when the global "done" was
updated.

So, now I've tested MountainCar, CartPole, and AcroBot, all of which run
multiple episodes back to back just fine, but they don't seem to be
learning the task. Their performance is poor.

I need to make sure that they are getting all the proper inputs and
generating all the needed outputs, and/or maybe just give it more time
(but it really shouldn't take long). I also need to compare with
learning on the Open AI Gym version of CartPole ... I thought that it
learned that one prettty fast.