#34 Can recreate some original Picbreeder images
Thanks to some help from Joost, I was able to recreate a simple genotype from the original Picbreeder (he sent it to me, and the simple.xml file is being added in this commit).

To fix this, I had to change my sigmoid and gaussian functions (I actually kept the old ones too). I also had to add an option that does one last pass on the output layer of the network to read off any latent activations.

Although this works, none of the original Picbreeder genotypes are perfect. They are all closer, and some are very close, but most are way off. I think there are still node ordering problems ... will need to look into it.