#27 Slightly more efficient file saving and iteration tracking
I now save scores as txt files instead of saving the whole behavior
ArrayList as an xml file, which should be quicker since it is less file
io. I also modified how iterations are counted: Now, every individual
that is produced counts as an iteration, which means that crossover
leads to two iterations, which I think is the fair way of comparing with
previous work on innovation engines, and also means that the same number
of iterations will take less time ... albeity perhaps with the
consequence of there not being enough time to generate the best possible
results.