#36 MAP Elites now creates a gnuplot file
The plot file indicates how the fitness in each bin changes over time in a manner similar to what is in the original innovation engines paper. A few settings might be too specific to the innovation engine though, such as how many individuals constitute a generation. However, limiting the plotting in this way is necessary so that gnuplot's memory capabilities are not overwhelmed.