#32 Weird ImageNet results
I added some test images from the actual ImageNet data set (technically the tiny image net subset) to test GoogLeNet, and the results are weird. I was hoping that I would get some sort of clear distinction (success/fail) depending on whether or not I used preprocessing, but GoogLeNet seems to be bad on all of these images both with and without preprocessing. Makes me wonder about the DL4J implementation of this model, and the lack of a built-in preprocessor specific to GoogLeNet.