#32 Helper method: can't test directly
I made a helper method that converts a INDArray derived from an image back into a BufferedImage so that I can display it. This is useful, since it shows the effects of scaling the image. However, I cannot use it to show the effects of color mean subtraction, as done by the VGG16 pre-processor, since mean substraction creates illegal color values. Still, the helper method is useful.