#24 RL4J succeeds at RLGlue's CartPole
Increasing the Max Step parameter in the QLConfiguration constructor
allowed the network to learn a policy for CartPole that works so well,
it balances (ssemingly) indefinitely. This is good, since it indicates
learning is actually happening.

However: This means that RL Glue's usual cutoff for the maximum episode
length is being ignored, and it is likely that if a second episode (or
more) were conducted, that the starting conditions would be identical to
the first episode.