#24 Added back episode length restrictions for RL Glue domains
The RL Glue MDP now ends episodes if the number of steps exceeds a
threshold, as in my evolution task.

Incidentally, AcroBot was NOT able to learn successful behavior with RL
Glue, when the episode length restriction was in place (couldn't learn
it before either). Why is this task so hard? Can compare with Open AI
Gym performance?