#36 Evolve pitch/heading and set number of view angles
The pitch and heading now evolve ... not sure this is for the best, but
it is a nice option to have.

Also added setting for the number of angles the 3D shape is viewed for.

Also made change to innovation logging to set the eventual maximum for
the gnuplot file in advance (can see how much is left to go).

Only step left for shape innovation is making some batch files.