#33 Lots of cahnges to make DL4J network same as HyperNEAT network: WORKS!
The Unit test for TensorNetworkFromHyperNEATSpecification actually
works! This means that the HyperNEAT substrate specification for Tetris
is correctly decoded to produce a DL4J network, and then each weight
from the substrate network is correctly copied over to the DL4J network.

The only small problem left is that it only works if the substrate
network uses the ReLU activation function for both the hidden neurons
and the output. The output should actually be tanh or a regular sigmoid,
but my HyperNEAT networks don't have a way of specifying different
activation functions for different layers yet.

However, I did make a related change: there is now an option to make the
input neurons use the ID activation function, which is actually a
standard convention that I have ignored for years.

Other changed include: swapping of substrate width and height when
defining input shape for DL4J, and deletion of an unnecessary method in
TWEANN that had actually been replaced by one in ActivationFunctions.

NOTE: There was a bit of trouble getting the test to work in the context
of all other tests during Maven install, so I had to go back and tweak
substrate initialization and other issues with global variables that
never change in a normal run. I think I resolved the issue though ...