Completely changed how activation functinos work
I feel that the series of if statements used to choose between different activation functions was slowing networks down, so I instead made each activation function a class and put them all in a HashMap so the classes can be looked up quickly whenever an activation function is used. This passes all unit tests, but it is a big change. Hopefully nothing is broken.