#33 #24 Successfully evolve DL4J with HyperNEAT
The HyperNEAT to DL4J conversion process works well enough that I'm evolving a network to play tetris that is a DL4J network. Had to add a batch file, and add some special handling for population and archetype initialization, but it seems to work. Need to monitor and test.