#33 #24 Wider network test
The process width is now 8. However, the depth could not be increased: 4
is the maximum process depth allowed by a convolutional net in Tetris
that uses a kernal of 3 by 3 because the input width has shrunk so much
by the time it reaches the 4th layer (moving upwards).

This can be fixed by using zero padding, or allowing different layer
configurations at each level.