#33 DL4J networks not so fast?
I compared the execution time of DL4J to my TWEANN class, and the TWEANN
was actually much faster, consistently. This is unfortunate. I guess I
won't be using DL4J networks as much as I thought, but they are still
useful for RL/backprop, and it also might be easier to explore certain
types of architectures with DL4J than with TWEANNs, so I'll probably
still use them a lot, despite the inefficiency.

NOTE: They MIGHT be faster when GPUs are used ... I still need to check
this.