#24 Looks like swapping of height and width caused exception
Extensive test code tweaning eventually revealed that the only
difference between the network that behaved and the one that didn't was
that height and width were swapped for the final pass from conv to
output layer. Swapping these indexes fixes this, and unit tests still
pass. Still need to test backprop though.