Found actual Pogamut repositories
I was finally able to remove all Pogamut jars from the local repository
that is in GitHub. Instead, the pom file designates the version 3.3.0
pogamut repositories.

HOWEVER, when looking into this I noticed that 3.7 exists, and it seems
to add some really cool features! Specifically, much better navigation.

Anyway, the reason I was mucking about with this was to fix #41 but I
have no idea if this change actually helps anything.

Also, it might be worth while to look at #1 again now that new libraries
are being downloaded.