#11 Trouble generalizing GVGAI code
Introduced this class to be a universal parent class for my GVGAI
players. Mainly, I wanted to fix the bad use of static variables.
However, I now see that Darwin originally did this because the
high-level method for running a game evaluation loads everything from
file based on Strings, such as the level, the player agent class, etc.

I need to change this code to accept class instances of all of these
components instead.