#11 #20 Lots of refactoring of GVGAI code
Some bad design practices were fixed, such as loading of data from files
and class names over and over. Classes are (mostly) created once and
reused. Also, a static variable for storing the controlling network was
removed and replaced with a member variable in a parent class. Work can
now progress on the reactive agent, though there is a lot of work to do
there.

Additionally, a new helper method (which might need to be moved) runs a
game based on a String array description of the level instead of a level
file name. This should make encoding levels with CPPNs easier.