#9 Calculated loss derivatives
I added calculations for derivatives from the original Gatys et al paper
(https://arxiv.org/pdf/1508.06576.pdf) since this seemed to be what Alex
Black was pointing out in his suggestion. However, the code crashes with
an exception when calculating the style derivative. There is a problem
with the matrix multiplication.