#20 Basic evolution of levels possible, but can't be played yet
Levels can be interactively evolved with a tiny image preview, but they
cannot be played yet. Also, I think the sprite placement is still a bit
wonky.

Additionally, this only works for zelda and blacksmoke, and I will
either have to figure out how to generalize this, or add unique
character sets for each game that I generate levels for.