#20 Levels can be played!
Had to play the game in a separate thread to make it work.

The only issues now are to add support for more games, and add constraints to make the levels more playable. Currently, it is often impossible to beat a level.

Some of the Zelda results were cool, but none of the blacksmoke levels really were. Need to try more games