Cannot load prototype generator
Adam trained a Mario level GAN with Keras 2 which is included in this
commit, along with an attempt to import it into DL4J. However, the
version of DL4J I am using does not support Keras 2, only 1.1.

To make this work, I need the latest version of DL4J, but that is not
currently available in Maven, only from the master branch on GitHub.