Trying to generate a level by passing via options
I added an option to store a Level instance in the command line options,
and had to add lots of alternate methods to handle this type of level
initialization. I think those methods work, but that the level I'm
sending in is malformed.

I'm either not using the right code from Vanessa, or there is some
setting in that code that is not in mine.