Levels generated properly now
Part of the problem was that I did not realize that Vanessa added a
buffer to the start and end of each level for the opening flat section
and an ending victory point.

However, beyond this, there actually was a problem with the bottom of
the screen not being displayed. I believe I fixed it.