#50 Can generte all enemies now
Not sure this is the best way to do it, since there are so many enemy
types. Also, may want to restrict the total number of enemies, because
they are sort of overwhelming the screen in a really weird way.

Additionally, it looks like the levels generated by the CPPNs are not
completely deterministic, meaning that if I replay the same level
multiple times, it is slightly different. Very odd.