Adding back support for two Marios
I'm retracing some of the steps in the dagstuhl_play branch, but want to have an over all cleaner result that maintains the original A* agent, and makes it easy to turn two-player mode on and off.

So far, I've added the second agent, but need to add back the ability to interact between agents.

#48