#48 Many static elements now non-static
A lot of the original Mario code assumes there will only ever be one
Mario, and thus uses static methods that alter static variables.
However, with Luigi around, there needs to be a separate copy of certain
variables for each Mario instance, hence the changing of how these
methods are defined in order to allow each agent to be treated
differently.

Howevr, note that for evaluation scores, I think it is still assumed
that all evaluation stats correspond to Player 1 (this way, evolutino of
Mario agents is not broken)