#9 Add Python code for neural style transfer
Getting the neural style transfer algorithm to work in Java seems more and more like a fool's errand, as long as DL4J is so inferior to Tensorflow. However, I should be able to call a Python version of the neural style transfer algorithm directly from Java and get the result back. Therefore, I've copied such Python code over.

The only problem currently is that the program directly reads and writes image files to disk. I want to bypass this by passing the "image" in ... perhaps in json format, and then printing the output json to stdout and reading that back to interpret it as an image in Java.