#9 New python file takes multiple styles
I modified the original neural style command line processor to interpret the list of style inputs as style images that are to be applied separately to a given content image. The example.txt provides a way of running this.

The content image will be an actual file on disk, but the styles will come from Java in the form of json, so getting this program to accept json is the next step.