#9 PythonNeuralStyleTransfer works!
Running the Java program PythonNeuralStyleTransfer will read a style image and send it to the Python program as a json String, which runs the neural style transfer algorithm and saves the file to disk.

Next step is to make the Python program print the image's json String to stdout instead, and read that from the Java program.