#9 Load VGG once in advance
The original Python code was loading VGG for each image processed. Now, I only load it once and reuse it. It is only a little faster though ... the main bottle neck is still the iterations of the algorithm itself. However, I still might be able to do something with the TensorFlow session.