Save frames of 2D animation
This is unrealted to the neural style branch. It loads a network for 2D animations and saves high quality version of each frame.

Annoyingly, it seems that upgrades to v3 of MM-NEAT make the old networks from the human subject study slightly incompatible, so I couldn't actually use this class from this code base. However, when I copied this code back to MM-NEATv2, I was able to generate the animations I need (not committed to that code ... just temporary)