#9 Don't recompute content features
In the Python Neural Style Algorithm, I was able to move some code around to avoid recomputing the VGG features for the content image over and over. Since the image is constant, and only the style images change, this provides a small savings.

The pause when loading new images is still noticable, but this change is an improvement.