#51 Subtract from space at end of level for progress measure
The fitness function I was using was intended to add in time if the agent beat the level, but I was checking this wrong, since I assumed the distance traveled would equal the level length upon successful completion. However, this is not true, so I have subtracted a constant value from the level length which I determined empirically, but which I hope applies universally.