Allow determinism and turn off small networks for HybrID
We are trying to troubleshoot HybrID, so I turned off the smallerGenotypes option when the HybrID switch occurs. However, I'm still not certain that HybrID works. However, it should be easier to test this if we use determinisic evaluations. So, I added support for this in Tetris. I also added a batch file that runs deterministic evals in Tetris and swiches with HybrID.  This should give us a better sense of how/if fitness is changing at the HybrID switch, and also allows us to test with only one trial, which is quicker.

This is related to issues #18 and #67