Added scaling of input/output from CSV file
This greatly improves results, and the error on the polynomial test function is close to 0, but I'm not sure how close it has to be to be good enough. Sometimes results are off by a decimal point (order of magnitude), which is probably not good enough. Though, this might point to the inadequacy of MSE more than anything else.