#74 Normalize with respect to variance of inpus and outputs
Rather than strictly scaling input and target values based on the maximums, I am now scaling them by dividing by the variance. In theory, this should make all inputs and targets restricted to the range [-1,1], but there are some exceptions to this rule, which means that there is still a bug somewhere. However, the model evolves quite well despite this. Still, I need to fix this