Rename SZ Tetris batch dir
I don't want this directory to be named as if I have current or pending paper results yet, so I renamed it. However, this will probably cause merge conflicts when @leg2015 eventually submits a Pull Request to merge the work from dev_tetris. We'll deal with that when the time comes. In dev_tetris, we should update from dev and then also move the batch files.