Fixed assertion error preventing post evals
This update could be relevant to @leg2015 and @DevonFulcher . The post
eval batch file includes the -ea flag, meanign that assertions are
evaluated. This was causing Tetris runs using hand-designed features to
crash when an assertion was checked due to an ArrayIndexOutOfBounds
error, but there wasn't actually a problem. The assertion was simply
assuming there were more CPPN inputs than were actually needed. I
updated the assertions, so post evals on these experiments are now
possible.