Fixed basic UT2004 evolution!
Close #1

It turns out the problem was that some old utility class that Igor made
for collecting game state info for the bot was trying to store
information about Items in a Map, but the ItemType class is not
Comparable, and the code was using a TreeMap. I simply changed these
instances to HashMaps. I'm very confused as to how this error crept in
... presumably, the code worked at some point, but that was a long time
ago.

In any case, closing this issue sets up @friedmaa nicely to work on
UT2004.