Pogamut files from local repository
The Maven repository for Pogamut files seems to be down. I don't know if
this is a temporary or permanent issue, but it is preventing any work
from getting done on the code. Therefore, I simply added the necessary
pogamut libraries to the local repository, which works for now. I might
change it back later, preferably when upgrading to the current version
of pogamut, but that probably won't happen any time soon.