Massive update to Pogamut 3.7.0
My code was using Pogamut 3.3.0, and was using local jar files for
Pogamut libraries. Because of all the problems that @friedmaa and I were
having running Pogamut, I decided to bite the bullet and simply upgrade
the Pogamut version in MM-NEAT to 3.7.0. This means we can rely on
Pogamut libraries from Maven again, use the latest features, and also
the example bots. We need to be on the look out for errors from this
upgrade ... we'll keep all of these changes in the dev_ut2004 branch
until we are comfortable with the upgrade.

Note that this upgrade also adds some of the basic Pogamut archetype
bots. Specifically, a CTF bot was added (very relevant to @friedmaa ),
and I also replaced the existing versions of HunterBot and NavigationBot
with the new versions.