Commented ActivationFunctionMutationTest, edited the test #110
The original test was asserting that the two lists of activation functions would be different. However, it is possible that the mutated function in the mutated genome statys the same. This is because when a activation function is mutated, a new function is chosen randomly from the list of possible functions. The original function has a chance of being chosen. The new test asserts that there is at most one difference between the two genomes to account for the instances where the mutation does not change the genome.