Allowing for generalizable tetris convolutional architecture
The goal is to allow for redefining the triples in
getNetworkHiddenArchitecture() and the constants in
getNetworkConnectivity to form any number of convolutional layers
followed by any number of fulling connected layers. Tests could be run
on a different sequence of conv and full layers  with a simple code
change, but I do not think this would be fruitful. #16