#136 #21 Temporarily hard-code non-PO mode
For now, we can just manually tell PacMan not to use Partial
Observability. I just made this change, and now we are getting a
different error message (progress!). In particular, it is one of the
UnsupportedOperationExceptions, which is good, since we can easily
disable the use of these methods when we are using the different version
of PacMan. @Squillyprice01