Added basic ray-tracing bot from Pogamut
Since the problem with UT^2 in Team Deathmatch seems to be related to ray traces #138 I've added this basic ray-tracing bot from Pogamut. @friedmaa test to see if this bot can run in a team death match server. If not, then there just might be a bug with ray traces that makes them impossible to use (unlikely). I expect that this bot will work, and that we can then see how it uses raytraces, and hopefully change UT^2 to use raytraces in the same way.