Further upgrade issues to DL4J 1.0.0-beta
I updated the model zoo version as well, which caused lots of issues
with the innovation engine and Image Net code.

I also found some code from when I attempted to import a Keras model for
Mario GAN. This code never worked, so I simply removed it.

There was also code for Neural Style Transfer that someone wrote in Java
and I imported, and I tried to fix it as best as I could.