Two versions of GameBots
There is now a regular version of the game bots mod (that makes it easier to include the debugging nav grid etc) and a botprize version (which is required for botprize testing).

The user will need to copy the one they want to use to the UT2004\System dir, but at least both are here now.