NEEDS REVIEW Implemented RelativeNormalizationCharacterization
I have finished most of it but I need your help in fixing the error message. I am not sure if this is the way that you wanted it implemented. I had to add two methods to the BehaviorCharacterization interface which means I had to add a lot of unimplemented methods to all the classes that extend BehaviorCharacterization. Maybe there is a better way that you would want me to implement this?