#194 added a method forceGame in CustomExecutor and ExecutorFacade
forceGame takes a pacman controller, a MASController for the ghosts, a game, and a MOVE. All it does is advance the game one timestep and moves PacMan with the MOVE you send. This method is meant to be used in conjunction with The dummy test ghosts. You can set their moves and pacmans move for testing purposes.