Refactored further to make most methods static
#193 @friedmaa I refactored the class further. I haven't tested it, but
everything that worked before should still work. This change makes most
of the important methods in the class static. The benefit of this
approach is that we will be able to call these methods from outside of
this class without the need to re-write or copy code. However, the
downside is the massive number of method parameters that now have to be
passed around in most cases.