Attempted to allow parallel evals in pacman
I thought that the main problem was that each separate eval thread did
not have its own input/output mediator. This is certainly a problem, but
changing this did not fix the exception I received. Also, this change
introduces some inefficient copying. If this doesn't allow for parallle
evals, then it is not worth keeping it around, so I will immediately
revert this commit in a moment. Still, I want to be able to access this
code in the future if I change my mind.

#213