#205 Can successfully launch botprize server
I think this issue may be complete, though we need to see if it works on
another computer first. The code will now actually copy the version of
GameBots (which contains BotPrize) directly to the UT2004\System
directory, replacing whatever version is there. In addition, it copies a
GameBots.ini file that disables all visualization settings, so that
there is not cheating at the BotPrize game. Let's see if it works on
other machines, and then close this issue @friedmaa