OldActionWrapper can create AgentMemory and AgentBody
Static methods for creating these data structres from a bot instance are now available, which are needed to send as parameters to many of the old-style bot Actions.

However, this also caused some changes in the type of PathPlanner used by UT^2, so we need to check and make sure that UT^2 still works appropriately.

Note also that the way the AgentMemory constructs its PathPlanners may be problematic ... may need to create them once and save them somewhere for later.

#238