#248 #249 #250 #251 #252 #253 Big Push. Read Description!
#249
Created OldToNewGhostIntermediaryController. As of right now, it takes a SharedNNGhosts<T> as its constructor. This is given to it in the only place it is constructed, MsPacManTask. The idea behind this class is that all of the decision making logic will be performed in this class. It contains four ZombieControllers and a MASController  made up of those four ZombieControllers. This class will set the moves of each of the zombies (one per ghost). When the MASController calls the getMove method, it will return those moves given to the Zombies by this class.

#250
MsPacManTask now constructs OldToNewGhostIntermediaryController correctly.

#251
Created POGhostsCheckEachDirectionMediator

#252
Created ZombieGhost.java. These are the ghosts that return moves set by the OldToNewGhostIntermediaryController. I am not sure how this code will ever interface with Piers competition code but this is the only way I could interface with our code.

#253
Added a classOption in Parameters.java called staticPacManPO. It builds a starter pacman to evolve against.

Altered MMNEAT to not crash when trying to evolve ghosts.

Added a constructor to GhostControllerFacade that takes an OldToNewGhostInterMediaryController.

Added methods getAction and getMove to NewGhostController in a similar fashion to NewPacManController. Because many classes extend NewGhostController, I added the unimplemented methods to each. getAction and getMove will never be called except by my new code, so this addition is backwards compatible. The files affect are:

AggressiveGhosts.java
BetterLegacy.java
SharedNNCheckEachDirectionGhostsController.java
ChaseFleeGhosts.java
DeterministicLegacy.java
RandomGhost.java
StarterGhosts.java
VeryRandomLegacy.java