#247 Testing method out: can evolve across multiple maps
The method that evolves across multiple maps seems to successfully run
one map and then the next. Tested it with the BorPrizeServer class to
run two maps back to back.

Note that this does not assure that the fitness/score results being
returned are correct, so this still needs to be verified @friedmaa