Fix in recurrent link visualization
TWEANN links are supposed to know whether or not they are recurrent, but
the boolean value used for this doesn't seem to really work. We should
probably remove it.

However, for now, I've just fixed the link visualization so that
recurrent links will be drawn in green. To determine if a link is
recurrent, I compare the indexes of the source and target neurons in the
archetype to see if the link goes forward or backward.

Note to @DevonFulcher regarading #259 . I evolved using the established
Tetris experiment batch files and did a post watch, and it turns out the
recurrent links were there all along.