#216 added eatPill method to GhostPredictionsFast
This method should be called when a pill is eaten. It updates the edible probabilities of all ghosts that we have info for atm. Essentiall, if we saw a ghost as a threat, rounded a corner into a pill, that ghost will be predicted as edible rather than a threat.