my implementation of the beginning of normalized nodes #162
I decided to sacrifice a little memory with the data members normalizedMemorySumVarianceTerms and normalizedMemorySum so that iterators and loops do not have to be used. My implementation closely follows the code snippet from the article and paper. I am not sure what will be done with epsilon. In the paper they say that "epsilon is a constant added to the mini-batch variance for numerical stability." Although I have not read the whole paper yet. Will epsilon be evolved like gamma and beta? The only thing that I changed to the old code is change the FullNodeGene constructor to protected so that I could override it.